+++ test/test-app.js
@@ -58,6 +58,8 @@ describe('generator-dizmo:app', function () {
                 'my-dizmo/gulp/tasks/document/watch.js',
                 'my-dizmo/gulp/tasks/lint',
                 'my-dizmo/gulp/tasks/lint/index.js',
+                'my-dizmo/gulp/tasks/outdated',
+                'my-dizmo/gulp/tasks/outdated/index.js',
                 'my-dizmo/gulp/tasks/upload',
                 'my-dizmo/gulp/tasks/upload/index.js',
                 'my-dizmo/gulp/tasks/watch',
@@ -178,6 +180,7 @@ describe('generator-dizmo:app', function () {
                     'url': ''
                 },
                 'scripts': {
+                    'prebuild': 'node ./gulp/tools/run-task.js outdated',
                     'build': 'node ./gulp/tools/run-task.js',
                     'clean': 'node ./gulp/tools/run-task.js clean',
                     'deploy': 'node ./gulp/tools/run-task.js deploy',