+++ test/test-sub-coffeescript.js
@@ -60,6 +60,8 @@ describe('generator-dizmo:sub-coffeescript', function () {
                 'my-dizmo/gulp/tasks/document/watch.js',
                 'my-dizmo/gulp/tasks/lint',
                 'my-dizmo/gulp/tasks/lint/index.js',
+                'my-dizmo/gulp/tasks/outdated',
+                'my-dizmo/gulp/tasks/outdated/index.js',
                 'my-dizmo/gulp/tasks/upload',
                 'my-dizmo/gulp/tasks/upload/index.js',
                 'my-dizmo/gulp/tasks/watch',
@@ -186,6 +188,7 @@ describe('generator-dizmo:sub-coffeescript', function () {
                     'url': ''
                 },
                 'scripts': {
+                    'prebuild': 'node ./gulp/tools/run-task.js outdated',
                     'build': 'node ./gulp/tools/run-task.js',
                     'clean': 'node ./gulp/tools/run-task.js clean',
                     'deploy': 'node ./gulp/tools/run-task.js deploy',