+++ test/test-sub-typescript.js
@@ -59,6 +59,8 @@ describe('generator-dizmo:sub-typescript', function () {
                 'my-dizmo/gulp/tasks/document/watch.js',
                 'my-dizmo/gulp/tasks/lint',
                 'my-dizmo/gulp/tasks/lint/index.js',
+                'my-dizmo/gulp/tasks/outdated',
+                'my-dizmo/gulp/tasks/outdated/index.js',
                 'my-dizmo/gulp/tasks/upload',
                 'my-dizmo/gulp/tasks/upload/index.js',
                 'my-dizmo/gulp/tasks/watch',
@@ -189,6 +191,7 @@ describe('generator-dizmo:sub-typescript', function () {
                     'url': ''
                 },
                 'scripts': {
+                    'prebuild': 'node ./gulp/tools/run-task.js outdated',
                     'build': 'node ./gulp/tools/run-task.js',
                     'clean': 'node ./gulp/tools/run-task.js clean',
                     'deploy': 'node ./gulp/tools/run-task.js deploy',